+++ CHANGELOG.md
@@ -20,6 +20,18 @@ For more information on SemVer, please visit http://semver.org/.
 
 ---
 
+
+## Version 1.4.15
+### 2014-01-31
+
+#### Bugs
+
+* Fixed static declaration
+
+#### Experimental
+
+* Added `WPDKScripts` class
+
 ## Version 1.4.14
 ### 2014-01-29